// src/type/person.types.ts
export interface IPerson {
    firstName: string;
    lastName: string;
    fatherName: string;
    identificationNumber: string;
    birthPlace: string;
    birthDate: string;
    certificateIssuanceCity: string;
    nationalCode: string;
    address: string;
    phoneNumber: string;
    postalCode: string;
    share: number | null;
    shareType: string;
    businessId?: string;
}